moves HomePageView context into liiweb
