+++ liiweb/views/general.py
@@ -1,8 +1,18 @@
 from django.views.generic import TemplateView
 
+from peachjam.models import Judgment, Legislation
+
 
 class HomePageView(TemplateView):
     template_name = "liiweb/home.html"
+
+    def get_context_data(self, **kwargs):
+        context = super().get_context_data(**kwargs)
+        context["recent_judgments"] = Judgment.objects.order_by("-date")[:5]
+        context["recent_legislation"] = Legislation.objects.filter(
+            metadata_json__stub=False
+        ).order_by("-date")[:10]
+        return context
 
 
 class PocketlawView(TemplateView):
